requirements.txt
# Python Dependencies for SyndraShell
# Only actively used dependencies

# Fabric - Main framework
python-fabric-git

# Core dependencies (ACTIVELY USED)
PyGObject>=3.42.0          # GTK3 bindings for UI
Pillow>=9.5.0              # Image processing for wallpaper thumbnails
pywayland>=0.4.14          # Wayland protocol bindings
setproctitle>=1.3.2        # Process title management
watchdog>=2.3.1            # File system monitoring for wallpaper directory

# Logging
loguru>=0.6.0              # Logging framework

# Reserved for future features (not currently used)
# ijson>=3.1.4             # JSON parsing
# numpy>=1.24.0            # Numerical operations
# psutil>=5.9.0            # System monitoring
# requests>=2.28.0         # HTTP requests
# toml>=0.10.2             # TOML configuration

